FaceDetectionApp/requirements.txt
absl-py==0.12.0
cachetools==4.2.1
camera==1.3.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
cycler==0.10.0
Cython==0.29.22
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
idna==2.5
importlib-metadata==3.10.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
pandas==1.2.3
Pillow==7.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.11.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
torch==1.8.1+cu102
torchaudio==0.8.1
torchvision==0.9.1+cu102
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.21.1
Werkzeug==1.0.1
zipp==3.4.1